testing/anbox: upgrade to 20200408, drop lxc dependency and fix init script
drop lxc dependency, as apk detects the lxc-libs dep and the init
script creates the required directory itself.

fix the init script: when the container manager failed, the openRC
service wasn't able to start. It could not be started again because
anbox-bridge.sh errored out: it was unable to create an already
existing interface.