main/mariadb: set CMake build type to MinSizeRel
It would be desirable to compile mariadb with the None build type to
make it pick up our CFLAGS/CXXFLAGS from /etc/abuild.conf.
Unfortunately, mariadb fails to build with the None build type. For this
reason, we use the MinSizeRel build type as a compromise as it is closet
to our default /etc/abuild.conf (i.e. uses -Os).