community/py3-flake8: fix pyflakes incompatibility
flake8 refuses to start with pyflakes 2.2.x. Work around this, without
downgrading pyflakes, by shipping the current master of flake8 and the
unmerged patch that makes it compatible with 2.2.x. I've tried to apply
the patch on top of the latest release, but that did not work. This
version appears to be working fine however, and I think it's the best we
can do in this situation. I'll add a warning to py3-pyflakes to prevent
this in the future.

Related: https://gitlab.com/pycqa/flake8/-/merge_requests/417
Fixes: #11378