main/py3-chardet: add check() & missing dependency
Adds a patch so that the tests work with pytest>=4.0. Upstream have
merged the patch but have not released a new version.

Before this change:

    $ sudo apk add py3-chardet
    ✂
    $ chardetect --version
    Traceback (most recent call last):
      File "/usr/bin/chardetect", line 6, in <module>
        from pkg_resources import load_entry_point
    ModuleNotFoundError: No module named 'pkg_resources'

After this change:

    $ sudo apk add py3-chardet
    ✂
    $ chardetect --version
    chardetect 3.0.4

The above commands were run inside docker-abuild with `dabuild sh`.

The change is required because `/usr/bin/chardetect` includes the line:

    from pkg_resources import load_entry_point

That means that py3-setuptools is required at runtime:

    $ python3 -c 'import pkg_resources; print(pkg_resources.__file__)'
    /usr/lib/python3.8/site-packages/pkg_resources/__init__.py
    $ apk info -q -W /usr/lib/python3.8/site-packages/pkg_resources/__init__.py
    py3-setuptools