community/lua-lgi: Remove bashisms to enable build
Lua-lgi package was using some illegal bashisms which should prevent
building. There is a commit hook, but somehow it was managed to be
skipped and thus the package was not properly build any longer.

What is even more worry some, is that these changes did not get
automatically get caught.

For example, when installing lua-lgi and lua5.1, the lua5.1-lgi package
is supposed to be installed. By removing the bashisms we can now
properly build lua-lgi again.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>