main/xrdb: don't depend on mcpp, use cpp from gcc instead
This dependency is entirely optional and thus not added to $depends.
The mcpp dependency was also entirely optional and should not have been
in $depends in the first place.

The xrdb package was the only package using mcpp, switching it to cpp
from the gcc package allows removing the mcpp package.