community/salt: fix run-time dependencies
tornado has been vendored in, so patches fixing tornado imports are no
longer necesarry.

Upstream switched back from pycryptodome to pycrypto due to issues with
some distro's, but still recommend it for security, so we patch it to
use pycryptodome.

msgpack was limited to <1.0.0. According to a commend this was [to
prevent 1.0.0rc1 being used][0], but that has now been released.

[0]: https://github.com/saltstack/salt/commit/e4ee32f2a068c9a90eb03a354f86a7f5b83c191c