main/ncurses: fix handling of globs and quote paths
Fix the handling of terminology* and st-* globs. They will be expanded
with $(find ...) and may return a list of files, so we need to loop over
the result.

Also do minor cleanups like quote variables with paths and keep the
/usr/lib/terminfo symlink.

fixes commit 95aa0e7217d (main/ncurses: upgrade to 6.2_p20200314).