community/h2o: fix dependencies and declare no check
Perl is required for the /usr/share/h2o/start_server script, which
execs perl. OpenSSL is required for running h2o as HTTPS server.

Declare the package !check for no since the checks are broken with the
current version of the package. They should work again once upstream
releases a new version.