community/sdl2: build the static library
the package featured a `-static` subpackage, but the subpackage was only
featuring `/usr/lib/libSDL2main.a`, which is a dummy file:

	> ar t /usr/lib/libSDL2main.a
	SDL_dummy_main.c.o

this commit make cmake build the static library, which makes the
subpackage include `/usr/lib/libSDL2.a` and work as intended.

I've tested it with an hello world program.