unmaintained/lizardfs: move from testing
Doesn't have a maintainer and still requires python2. Current git HEAD
has support for python3 [0]. Unfortunately, there hasn't been a new
release yet and upgrading to git HEAD will be a lot of work as various
file paths changed. I attempt such an upgrade but as I do not use the
software and the changes were too intrusive I decided against merging
them and moving this package to unmaintained instead [1].

If you use this software feel free to take over maintainership, update
it to a version which doesn't require python2, and put it back into
testing.

[0]: https://github.com/lizardfs/lizardfs/commit/10b4ab6c6ff791f3653f53d02beb52ee994c2f33
[1]: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/8591