community/qemu: run pre-install as pre-upgrade hook so users are created during upgrade
We introduced a new user in qemu previously with 2f32bfc751bae19369fc06666e8f4b2ce88d17fd,
but that user is only ever created when qemu is installed from scratch
and not when upgrading an exisiting qemu installation. This broke
upgrading from 3.11 -> 3.12 where we expect the qemu user to be available
in 3.12 but it wasn't unless you removed and installed qemu again.
Also redirect the stderr of adduser to /dev/null so users don't see
"adduser: user 'qemu' in use" during upgrades.

fixes #11618