main/ca-certificates: don't delete ca-certificates.crt
The ca-certificates.crt was previously not shipped with any package, but
generated from trigger script, so the post-deinstall needed to delete
it. Since commit 7363758e1d81 (main/ca-certificates: ditch python dep,
symlink ca-certificates.crt) this is no longer the case, and we ship a
pre-generated bundle in the ca-certificates-bundle subpackage. So we
should no longer delte ca-certificates.crt from post-deinstall.

fixes #11663