main/fontconfig: tests: -j1, cat log on failure
One of the tests, "run-test.sh", is currently failing on
build-edge-armhf. I was not able to reproduce it with an arm machine
locally. Using -j1 should hopefully fix it, and if it does not, at least
we should see more details about why it fails on the build machine.

Notably the gitlab CI job for build-armv7 also failed without -j1.