community/vlc: add Pulse support
This makes it depend on runtime at so:libpulse.so.0. Do note that this
does not mean you install the entirety of PulseAudio if you install VLC,
it only installs the libpulse package which does nothing when you're not
running the rest of PulseAudio. VLC can still play through Alsa just
fine.

On my PulseAudio system this fixes an issue where playback is really
choppy and "[00005580c7431380] alsa audio output error: cannot estimate
delay: I/O error" is printed to the console