main/cyrus-sasl: move plain back to libsasl package
The PLAIN auth mechanism is part of core SASL specification, and any
conformant client is expected to support this.  Splitting out the
PLAIN auth mechanism into a subpackage results in software breaking
until it is manually installed.  In essence, the PLAIN plugin shouldn't
have ever been split from libsasl but other plugins are fine to split.

Closes: #11739