community/py3-jsonpatch: move from testing
Move py3-jsonpatch and py3-jsonpointer from testing to community.

Take over maintainership of both packages - they only appear to be used by
cloud-init and it makes sense for them to have the same maintainer as
cloud-init. I emailed the current maintainer (@TBK) 8 days ago suggesting this
but received no response.

Align archs of the 2 packages to be the same as cloud-init.