community/py3-cx_freeze: fix permissions
Files in the package are set to 0700 and directories to 2700. These permissions
end up in the package, which prevents any user other than root to read these
files. Fix this by setting all files to 0644 and all directories to 0755.

This does not interfere with any executable scripts in this package, because
they still end up with executable permissions in the package.