testing/newlib: Re-add libc_nano.a
Due to some invalid shell commands in the APKBUILD file, we stopped
renaming the libc_nano and friends and thus missed those in our final
package, effectively breaking the package's function.

So we stop doing bash substitution, and no longer rely on unsupported
flags to read, making the line simpler to boot. Finally a period got
shifted for some reason or another resulting that overal, we did not
have our libraries any more.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>