main/vala: disable tests on ppc64le, fail to compile
A fix is available upstream: https://gitlab.gnome.org/GNOME/vala/-/issues/1068.
We can't just apply it though because it modifies the parts of Vala that
are written in Vala, so we'd have to rebootstrap the compiler. Since
a new release will be done soon for GNOME 3.38, we can just wait for that
and re-enable the tests when bumping.