community/prometheus: fix provided ldflags
`-extldflags` is used to provide LDFLAGS to the external linker.

159018314b (community/prometheus: fix provided ldflags, 2020-09-19)
extended LDFLAGS with flags that are meant for Go itself. As long as
`$LDLFAGS` is set to a value this works.

But when `$LDFLAGS` is empty, it would pass the flags ment for Go to the
external linker, which will then error with a usage message.

Fix this by only supply -extldflags when `$LDFLAGS` is not empty.