main/libsndfile: add dependency on opus
From the official libsndfile-1.0.29-r0.log:

	checking for opus >= 1.1 ... no

	configure: WARNING: *** One or more of the external libraries (ie libflac, libogg and
	configure: WARNING: *** libvorbis) is either missing (possibly only the development
	configure: WARNING: *** headers) or is of an unsupported version.
	configure: WARNING: ***
	configure: WARNING: *** Unfortunately, for ease of maintenance, the external libs
	configure: WARNING: *** are an all or nothing affair.

libsndfile seems to require that all external libraries (including opus)
are available. If this is not the case none of the audio formats
supported through external libraries actually work. This bug was
discovered by debugging mumble, which uses libsndfile and attempts to
play vorbis samples using it. As vorbis is supported through an external
library in libsndfile, this caused mumble to be incapable of playing any
sound files.

Fixes #11933

See also: https://github.com/mumble-voip/mumble/issues/4492