community/zabbix: fix setup subpackage
The setup package should contain the database schema's, but only
contained Makefiles. The reason is that `database` directory was copied
from `$builddir`, but the source is copied for each database, where the
actual build is done.

So copy the database schema's from `$builddir-postgresql`, like is done
for other subpackages, and make sure the `Makefiles` don't end up in the
package.

Fixes #11987