main/ndisc6: move from community
In an ipv6-only slaac setup, there is no DHCP server to provide DNS
servers to use. rdnssd, part of ndisc6[0] is a userspace program that
collects this information from router advertisements and updates
resolv.conf.

By having this in main, this can be available in the base setup, so that
Alpine Linux can be installed in an IPV6-only setup.

[0]: https://www.remlab.net/ndisc6/