main/openrc: fix osclock init.d script to prevent warning
This patch adds a stub start() function to init.d/osclock to remove
OpenRC warnings that appear during boot. There is no need for the
start() function to actually do anything as the osclock script's sole
purpose is simply to meet the "clock" dependancy of other init.d
scripts.

This patch was submitted upstream as PR 377.