main/linux-rpi: compile i2c-bcm2835 into kernel
A previous change, !12317, compiled the DS1307 RTC module into the
kernel so that Raspberry Pis with RTCs fitted could automatically have
the system clock set from the RTC by the kernel very early in the boot
process when the kernel is loaded, before any initramfs or root
filesystem is accessed.

However as the DS1307 is an I2C-based RTC, on Raspberry Pis it cannot
be used until after the I2C driver for the BCM2835 is first loaded. This
patch adds the BCM2835 I2C driver into the kernel so that it and the
DS1307 driver can both be used to actually achieve the result originally
intended by the !12317 change.