community/cloud-utils: make -growpart subpackage depend on partx
Partx was moved out of the main util-linux and into a subpackage
therefore cloud-utils-growpart's dependency for util-linux needs
to be changed to partx instead.

Also remove cloud-utils-growpart dependency on coreutils and add
patch growpart to use a flock command-line option that is understood
by both util-linux and bosybox's flock command.