main/unbound: remove dependency on dns-root-hints
The dns-root-hints dependency has been introduced in [1]. Strictly
speaking dns-root-hints is not necessary as unbound provides builtin
root hints. This change switches to using the builtin root hints by
default, thereby avoiding installation of a cron and a dependency on
gnupg. If desired, users can still manually install and configure
dns-root-hints with unbound.

See #11324 for further information on this topic.

[1]: https://github.com/alpinelinux/aports/pull/5950