community/py3-hyperopt: fix checksum
py3-hyperopt was upgraded to 0.25 in September
(3057b809767a3e2824dda731f8583198d0cb1784) in October upstream created a
new tag for the 0.2.5 release, overwriting the existing one. Thereby
causing a new tarball, with a different checksum, to be generated.

Bump pkgrel for changes potentially made in the new release.