community/dolphin-emu: disable llvm support to fix build
Since LLVM is not explicitly listed in $makedepends I am under the
impression that it was not intended to be supported and was just pulled
in indirectly by accident through some existing dependency. Disabling
LLVM fixes the following build failure:

	CMake Error at /usr/lib/llvm10/lib/cmake/llvm/LLVMExports.cmake:1276 (message):
	  The imported target "LLVMDemangle" references the file

	     "/usr/lib/llvm10/lib/libLLVMDemangle.a"

	  but this file does not exist.

Alternative solution: Add llvm10-dev and llvm10-static to $makedepends.