community/py3-typed-ast: fix linking against wrong lib
Due to symbol coliissions with python, some parts are incorrectly
linked against python instead of typed-ast. This causes struct layout
mismatches, which specifically impacted s390x, due to being big-endian.

This was uncovered by the black test suite, which resulted in an error
message:

```
SyntaxError: can't assign to ()
```

Upstream provided a [patch][0] to mitigate this issue.

[0]:https://github.com/python/typed_ast/pull/152