community/msitools: add patch for set_property
This is a feature added in master, but not part of a release yet. It's
required if you want to add custom post-install actions, so bacport it.

To make the build work, we have to make sure the autoconf/automake files
are up-to-date. The `config.vapi` file that is required is missing as
well, but has been added in master, so we take that (with one missing
field added).