community/postfixadmin: config.local.php symlink missing
When setup.php does a sanity check of the config it checks the config
via the symlink /usr/share/webapps/postfixadmin/config.inc.php i.e.
not directly /etc/postfixadmin/config.inc.php, this works as we have
this symlink in place.

But it does not find the optional/recommended config.local.php, resulting in an
incorrect warning:
 Warning: config.local.php - NOT FOUND

Add a symlink for config.local.php as well.