community/dmd: update to 2.095.0 & use libunwind
The vfork() definition in the patch has been upstreamed in v2.095.0,
so the diff can be removed.
This also brings in the support for libunwind which is in the process
of being upstreamed (the first few commits are already in master),
and is necessary as libexecinfo is quite buggy and leads to a lot
of test failures/crashes.
The upstream code supports both llvm-libunwind and libunwind,
but for the sake of coverage this goes with the former,
as the later will probably be covered by LDC.