main/dtc: add py3-libfdt subpackage
pylibfdt is included in the dtc sources and allows using libfdt
from Python.

For some reason, the pylibfdt Makefile ignores $(DESTDIR) at the
moment. Add a small patch to fix this so the Python library is
installed properly. This is just temporary, the next release will
introduce Meson build scripts which do not seem to have that problem.