main/nginx: change default dir for vhost configs from conf.d to http.d
It was a mistake (and I admit it, it was my mistake) to start using conf.d
for vhost configs. This directory is better suited for configs to be
included in the root context, not in the http context. And it's even more
evident since the introduction of the stream module - where should be
stream servers defined? I think that the most intuitive location is a
stream.d directory. Similarly, http servers should be in http.d directory.

The new layout for Alpine is the following:

- nginx.conf - the main nginx config that includes config snippets
- conf.d/*.conf - configs included in the root context (will be added later)
- http.d/*.conf - configs included in the http context (new, former conf.d)
- modules/*.conf - configs for loading modules
- stream.d/*.conf - configs included in the stream context (new)