main/fish: make util-linux a dep of fish-doc and add a README
It is required to have util-linux to use the `help` command in any of
the builtins fish provides as the internal `__fish_print_help`  function
uses the `ul` program that `util-linux` provides.

`__fish_print_help` also returns early if there are no manpages
installed which are present in `fish-doc` so can we can kludingly deal
with this by tying the `util-linux` dependency to the -doc subpackage.

This is not perfect as many users that want `fish-doc` don't want
`util-linux` but it is better than giving a free `util-linux` dependency
to anyone that installs `fish`