main/boost1.75: finish renaming process
This finishes renaming `boost` to `boost1.75`, this is a versioned
version of the package which can be installed side-by-side with another
versioned boost package.

Since this is the package that we currently build all our packages that
need boost against we make it also claim the unversioned names like
`boost-dev`, `boost-libs`, `boost-doc`, etc.