main/bonding: add support for bond-members statements
The newer ifupdown-ng bond executor uses bond-members instead of bond-slaves and
rewrites the term as support for legacy terms was reimplemented as a term rewriter
when namespaced terms was introduced in ifupdown-ng 0.10.

Alpine still uses the legacy bonding scripts for 3.13, so we need to handle the
rewritten term as well here.

Fixes #12319.