main/boost1.75: don't use custom statx implementation
When there is no statx function, but the statx syscall is detected,
boost-filesystem implements a statx function that directly calls the
syscall.

This is broken, because statx might not be implemented on the running
system, and there is no logic to fall back to fstatat.

Remove the build-time detection of the statx syscall, so that the code
falls back to `stat`.

Fixes #12349