main/fortify-headers: remove breakage of ppoll()
The fortify-headers/poll.h definitions incorrectly forward the
ppoll() function: on some archs (armv7 for instance), musl
renames the symbol to accommodate for time_t size, and
fortify-headers misses that.
This causes ppoll() to break on all time32 systems.
This commit axes the ppoll() redefinition. Fortifying ppoll()
is of very little benefit anyway.