community/gogs: modernize
With go modules, it's no longer needed to move the builddir as was
expected with GOPATH.

Also use chmod-clean, otherwise the module cannot be cleaned up
properly, which fails the build.

Finally, move GOPATH from `$startdir` to `$srcdir`, which is more
common, and does not polute the `$pkgdir` with go modules, as the names
overlap.