community/gitlab-runner: fix issue in docker dependency on mips
pkg/system/stat_linux.go in github.com/docker/docker assigns a 32-bits
(on mips64) Rdev value to a 64-bits field, which results in a build
error. docker upstream already fixed this in v20.10, but gitlab-runner
is still using an older version and upgrading it proofs problematic.

Just fix this issue in the vendored version of the lib.