community/prosody-filer: fix build
Package does not build because it could not find the toml dependency.
Even more, the APKBUILD just did a `go get` for that package, meaning
that the module is not even fixed.

Modernize the APKBUILD and use go modules to fix this issue.

Also consistently set GOPATH so that go clean -modcache works properly.