community/gpgme: ship static libraries
Adds the "--enable-static" flag to the APKBUILD to build gpgme's static
libs. This mirror the behaviour that exists in Debian's gpgme-dev
package.

This change would be useful for us as we're trying to statically link an
application using buildah, which has a hard requirement on gpgme. As
this package does not build the static libraries, it's not currently
possible to do this with Alpine, but it is possible with Debian.