community/collectd: split many plugins into subpackages
collectd includes dozens of plugins. I moved into subpackages all
plugins depend on any extra library (except libyajl), plugins for
programs that are not in the base Alpine installation, plugins for
monitoring hardware devices and some bigger plugins.

I also added meta-package collectd-plugins-all that can be used to
quickly install all the plugins.

This allowed to decrease the size of the collectd package from
2.2 MiB to 1.5 MiB and removed dependency on libcrypto1.1, libcurl,
and net-snmp-agent-libs.