community/pipewire: temporarily remove provides for -pulse
Currently this only seems to cause problems. On clean installs, "apk add
pulseaudio pulseaudio-alsa" would not install PulseAudio but
PipeWire-Pulse instead, which obviously doesn't make sense.
This would be less of a problem if -Pulse was actually auto-started, but
it currently requires manual intervention to start and thus effectively
breaks audio for people not even wanting PipeWire in the first place.

Anyhow, the current behaviour is broken so let's just revert to -Pulse
not providing PulseAudio for now while we work on a proper fix