testing/py3-weasyprint: rename aport to weasyprint
This is mainly a CLI tool, not a Python library, that just happens
to be written in Python. We don't use py3- prefix in such cases
(and neither other distros).
More precisely, the aport should be named weasyprint, but
/usr/lib/python3 should be splitted into a subpackage py3-weasyprint
(i.e. the origin package would contain just /usr/bin/weasyprint in
this case), but doing that now would break existing installations,
so let's postpone it to the future.