community/qemu: drop obsolete signal 33/64 patch
This was a patch used to allow compatibility with programs compiled with older
versions of Go.  For the past several years, Go has used a different signalling
mechanism between goroutines which no longer needs this patch.

A side effect of this patch is that pthread_cancel was broken, which caused
problems with running Qt applications under qemu emulation.