community/pass: add coreutils dependency
Pass uses the tr graph character class which isn't currently supported
in Busybox tr but is supported in GNU tr. The absence of a tr with graph
character class causes tr to use the literal characters '[:graph:]'
instead of a character class which causes pass to generate very weak
passwords. Add coreutils as a dependency so pass can use a tr
implementation with the graph character class support.

Fixes #12583