testing/libguestfs: upgrade to 1.44.1
This upgrade is required for compilation with OCaml 4.12 that
is now included in edge.

The python 3.8 patch has been reverted as it doesn't apply
and the build appears to succeed without it - we're also on
python 3.9.

I've included a new patch that ensures that everything compiles
correctly even if --disable-ocaml is selected, I'll try and get
this accepted upstream soon.

Additionally removed an unused variable and removed $pkgname
from the download string to satisfy the linter.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>