main/vim: fix shellpipe for ash and dash
vim uses a list of known shells to include stderr in the redirection for
shellpipe. Include busybox ash and dash to this list.

It would be better to use POSIX redirection by default, but that would
break backwards compatibility of default behavior.

Submitted upstream: https://github.com/vim/vim/pull/8180