community/cloud-init: upgrade to 21.2
Upgrade to latest version.
Remove CVE-2021-3429.patch as included in this version.
Add parted to depends as this version will use it in preference to blockdev.
Remove util-linux from depends as use of wipefs is undocumented upstream.
Do not include some, non-relevant to Alpine, cloud-init modules in package.
Update README.Alpine to add notes about dependencies.