community/sway: fix segfault on start
The build of sway statically linked in libudev-zero, but it also links
against wlroots, which links dynamically against eudev-libs. As a
result, a call to udev_monitor_filter_update() in eudev-libs calls
while iterating a list into libudev-zero for udev_list_entry_get_next(),
which expects a different list format and results in iterating past the
end of the list.

This adds eudev-dev as an explicit dependency, to make sure that
libudev-zero doesn't get used in the build.