community/easypki: fix build on riscv64/mips64
The original boltdb is archived, but does not support the architectures
we need, like mips64 and riscv64.

A fork of boltdb, called bbolt, has support for these arches, but it's a
different module, which makes it harder to switch to it, and the quality
of bbolt is questionable.

Instead, use a fork of bolt, which just adds support for these arches,
which are trivial files with some constants.