testing/yubico-piv-tool: move check-dev to makedepends
The configure script fails if check-dev is not installed, even if tests
are not executed. As such, it is technically an unconditional build
dependencies. That is, with options="!check" (as on the riscv64
builders) this aport previously refused to compile. By ensuring that
check-dev is always installed we can fix this compilation error.

Note that it will not be picked up automatically as a run-time
dependency, thus this shouldn't cause any issues.