community/gitlab-runner: update fix for riscv64
The old replacement of github.com/kr/pty only replaced v1.1.1. But in
the latest version of gitlab-runner, v1.1.5 is being used, so it is not
replaced anymore.

Instead of replacing a specific version, replace all versions. And
instead of using a patch, use `go mod edit` to add the replace
decleration.