Changes my name in packages I contributed to
I dont know why, but when i started to contributed i kind of used my
nick name and not my real name. I want to change that. So I'm not sure
if I should split this in individual commits or just use something like
this. So, please see this more like a question how i can do things right
:)