testing/fastd: move cmocka-dev to makedepends
The meson.build explicitly checks for this dependency and refuses to
continue if it is not installed:

	test/meson.build:5:0: ERROR: Dependency "cmocka" not found, tried pkgconfig

However, since checkdepends are not installed with options="!check" the
build will always fail if this option is set. For example, on the
riscv64 builders. Moving the dependency to makedepends fixes the build
on riscv64. Keep in mind, that compile-time dependencies do not
automatically cause new run-time dependencies, thus this shouldn't be an
issue.