testing/crypto++: explicitly build static library too
Since a -static subpackage is provided, I assume the intend here is to
also build a static library. However, presently only a dynamic one is
created in the build() stage. Afterwards, the static one is created as a
byproduct in the check() stage due to $LINK_LIBRARY (which defaults to
libcryptopp.a, not libcryptopp.so) being a dependency of cryptest.exe
which is a prerequisite of the check target.

If the tests are not executed (e.g. options="!check"), creating a
-static subpackage will fail without this commit. For example, on the
riscv64 builder.