community/icingaweb2: added missing directory with correct permissions
When generated the config file for apache2, by default icingaweb2
does not add the env variable ICINGAWEB_STORAGEDIR, which in turn
is set as /var/lib/icingaweb2.

This patch fixes the error:

"The Icinga Web 2 storage directory defaults to "/var/lib/icingaweb2", if
not explicitly set in the environment variable "ICINGAWEB_STORAGEDIR".
The directory /var/lib/icingaweb2 does not exist."