main/gnupg: split into multiple subpackages
Most users and programs don't need a full GnuPG suite (with its zoo of
dependencies), but only some of its components. The goal is to reduce
number of installed dependencies and disk space.

The structure of subpackages is based on Debian's gnupg packages,
with a few differences in naming. At first I thought that it's
overcomplicated, but after understanding GnuPG components and how
are they related, I realized that it's actually reasonable and it's
the right approach.