community/ostree: build with libcurl and openssl instead of libsoup
libsoup is Gnome HTTP client/server library. It's a huge dependency
in terms of number of transitive dependencies and total size (for
example, it installs gnutls and sqlite). This is clearly not a good
choice for Alpine Linux.

Unlike libsoup, libcurl and openssl are common dependencies that most
users already have installed on the system. And it has more advantages,
mainly support of HTTP2.

For reference, Fedora switched from libsoup to libcurl + openssl with
version 2017.8 (Fedora 26).