community/libfreeaptx: move away from git version
This package was shipping 1 newer commit than the last release, but the
commit itself did nothing useful for us. So, let's switch back to the
release version. Besides the pkgver was using a weird date notation
anyways, this prevents confusion.

Also ship a single patch which fixes the version provided in the
Makefile