main/man-pages-posix: split from main/man-pages
The POSIX man pages are taken from IEEE Std 1003.1-2017 and as such are
under protected under the copyright of the Institute of Electrical and
Electronics Engineers (IEEE) and The Open Group. They are, contrary to
main/man-pages, not licensed under a FSF/OSI license. As such, it makes
sense to separate them from main/man-pages.