main/pcsc-lite: split off debug/spy library
The PCSC spy library shouldn't be used runtime, or instead of the normal
pcsc-library. It really is only used by pcsc-spy binary, which is a
debug/logging kind of spy tool.

Further more, the spy library (and binary) are GPLv3 licensed, contrary
to the BSD licenses for the other code.

This patch fixes all that.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>