main/openssl: disable SSE2 on 32-bit x86
Some SoCs, such as the Vortex86, have buggy CPUID code, which results in
programs using CPUID for detection to assume the presence of SSE2.  It
should be noted that the baseline architecture for 32-bit x86 is i586
at the moment, but whether MMX is required or not is unspecified.

It may be better to target 386 for OpenSSL on 32-bit x86.

But this should fix #13016.