main/tevent: move cmocka-dev from checkdepends to makedepends
The CMake build system checks for the presence of this dependency and
refuses to build if it is not installed. As such, this dependency needs
to be moved to $makedepends, otherwise building with options="!check"
(e.g. as configured on the riscv64 builder) will fail.

This fixes the build on riscv64.