community/jimtcl: backport fix for memory leak
While debugging an issue with linking OpenOCD dynamically against
libjim [1], I noticed that OpenOCD contains a workaround for a memory
leak in jimtcl which is enabled by default [2]. This workaround causes
an infinite loop when OpenOCD is linked dynamically against jimtcl. As
such, I disabled it in OpenOCD for now [3]. However, I think also makes
sense to backport the fix for the memory leak from upstream jimtcl.

[1]: https://gitlab.alpinelinux.org/alpine/aports/-/issues/13046#note_182639
[2]: https://github.com/openocd-org/openocd/blob/050fcb176071cadc7142c4d9acd3f5a9e67d3ac6/src/helper/command.c#L140-L170
[3]: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/26053
[4]: https://github.com/msteveb/jimtcl/pull/198