scripts/mkimage.sh: switch from abuild-apk to apk
The issue is that 'abuild-apk' creates root-owned files in the workdir,
and they cannot be removed later without root privileges.

'abuild-apk' is called only with 2 commands (add --initdb, update),
and in the context of the script they don't seem to be requiring
root privileges, so I'm proposing to substitute the calls with 'apk'.
This will automatically resolve the issue mentioned above.

Signed-off-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@bell-sw.com>