community/ghc: use embedded libffi, llvm12, drop gold
Since ghc depends on it self it becomes non-trivial to libffi with ABI
breaking changes. Doing so would require 2 different version of libffi
to be available at the same time (old version to make gch runtime work,
and new version to rebuild ghc with).

We simplify things by simply using the bundled libffi, so system libffi
is completely unrelated. The bundled libffi is built shared but is
installed under /usr/lib/ghc-* so we set ldpath so abuild can find it
and avoid pull in dependency of external libffi.

Make sure that we don't use ld.gold, which seems to be buggy.
https://gitlab.haskell.org/ghc/ghc/-/issues/17508

Pin llvm12 and ad patches from upstream to make the testsuite pass with
llvm12.

Skip some tests that are known to fail.

Replace the autoconf-ac-prog-c99-fix.patch patch with the corresponding
patches from upstream.

Drop build options:
 - INTEGER_LIBRARY=integer-gmp:
   upstream replaced integer-gmp with a unified ghc-bignum
   https://gitlab.haskell.org/ghc/ghc/-/merge_requests/2231

 - BeConservative, GhcStage3HcOpts, SplitSections:
   Those seems to have caused test suite to fail. Removing them made
   test suite pass.

Clean up the dev() function to use amove instead of cpio.