main/mosquitto: specify WITH_CJSON for install
When bootstrapping, if the `WITH_CJSON` make variable is not specified
to the `make install` command, it tries to build mosquitto again with
CJSON support, which fails because the headers are not available.

Specify the `WITH_CJSON` variable to the `make install` command as well
to prevent the rebuild from happening.