community/ruby-sdbm: new aport
This gem is unbundled by upstream Ruby from 3.0.0.

When the gem was bundled its pkgver was set to that of the ruby-lang
package, now its pkgver is set to its upstream version which is lower so
users upgrading need to pass `--available` to the `apk upgrade` command.

I could find no aport that depend on this gem.

Co-Authored-By: omni <omni+alpine@hack.org>