main/ruby-rake: new aport
This gem was bundled in main/ruby and it's on the list of the "Bundled gems"
(https://stdgems.org/#bundled-gems-ruby-302) for Ruby 3.0.2. I'm moving it
to a separate aport to provide a correct pkgver - to avoid confusion and
future complications if the upstream decide to remove it from the
"Bundled gems" (as they did for some other gems with 3.0.0 release).