main/sudo: backport patch to fix crash on old kernels
In postmarketOS (which uses Alpine's repositories to a large degree), we
unfortunately have many devices that use old kernels without the
getrandom() syscall. This backports a merged patch from upstream that
works around this by using sudo's own implementation of getentropy()
instead of the one from your C standard library as the default on Linux.