community/dtools: merge & upgrade to 2.098.0
There is no actual code change (dtools version is tied to community/dmd),
but we need to rebuild against LDC v1.28.0.
Since the lint rules could not be sanely satisfied with the old approach,
the 'dtools' package being a virtual package, all tools have been merged
into one. They are lightweight enough that the user will not see a difference.