community/gnome-shell: do not set cap_sys_nice
The way the current gnome installation works, setting cap_sys_nice
for gnome-shell makes it crash at startup. The reason is that
gnome-shell creates a dbus session on startup, which is only successful
if the process does not need to be treated securely (through
`getauxval(AT_SECURE)`). In consequence, setting some capability in the
binary, which is run under unprivileged gdm user, makes it impossible
to start the session.